# Performance improvements, serialization improvements, and added some tests