--- a/src/irsdkSharp.Calculation/IRatingExtensions.cs
+++ b/src/irsdkSharp.Calculation/IRatingExtensions.cs
@@ -5,6 +5,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Security.Cryptography.X509Certificates;
+using irsdkSharp.Models;
 
 namespace irsdkSharp.Calculation
 {
@@ -12,12 +13,12 @@ namespace irsdkSharp.Calculation
     {
         private static readonly double _initialConstant = 1600 / Math.Log(2);
               
-        public static Dictionary<int, double> CalculateIRatingGains(IRacingDataModel dataModel, IRacingSessionModel sessionModel)
+        public static Dictionary<int, double> CalculateIRatingGains(Session dataModel, IRacingSessionModel sessionModel)
         {
             if (sessionModel == null) return null;
 
             if (dataModel == null) return null;
-            var currentSessionId = dataModel.Data.SessionNum;
+            var currentSessionId = dataModel.SessionNum;
             var currentSession = sessionModel.SessionInfo.Sessions.Where(x=>x.SessionNum == currentSessionId).FirstOrDefault();
 
             if (currentSession == null) return null;
@@ -151,9 +152,9 @@ namespace irsdkSharp.Calculation
         {
             var sessionModel = racingSDK.GetSerializedSessionInfo();
 
-            var dataModel = racingSDK.GetSerializedData();
+            // var dataModel = racingSDK.GetSerializedData();
 
-            return CalculateIRatingGains(dataModel, sessionModel);
+            return CalculateIRatingGains(racingSDK.Session, sessionModel);
         }
     }
 }