--- a/src/irsdkSharp.Console/Program.cs
+++ b/src/irsdkSharp.Console/Program.cs
@@ -2,7 +2,6 @@
 using irsdkSharp.Serialization;
 using irsdkSharp.Serialization.Models.Session;
 using System;
-using System.Data.SqlTypes;
 using System.IO;
 using System.Linq;
 using System.Threading;
@@ -80,7 +79,6 @@ namespace irsdkSharp.ConsoleTest
                 // Check if we can find the sim
                 if (sdk.IsConnected())
                 {
-                    //Console.WriteLine("SDK is connected.");
                     if (!_IsConnected)
                     {
                         // If this is the first time, raise the Connected event
@@ -113,33 +111,34 @@ namespace irsdkSharp.ConsoleTest
                         _DriverId = (int)sdk.GetData("PlayerCarIdx");
                     }
 
-                    //var data = sdk.GetSerializedData();
+                    var data = sdk.GetSerializedData();
 
                     // Raise the TelemetryUpdated event and pass along the lap info and session time
                     //var telArgs = new TelemetryUpdatedEventArgs(new TelemetryInfo(sdk), time);
                     // this.RaiseEvent(OnTelemetryUpdated, telArgs);
 
-                    //if (data != null && _session != null)
-                    //{
-                      //Console.SetCursorPosition(0,0);
-                      //  var mm = IRacingSDK.GetMappedFile(sdk).CreateViewStream();
-                       // byte[] bytes = new byte[mm.Length];
-                       // var numBytes = mm.Read(bytes);
+                    // Is the session info updated?
+                    int newUpdate = sdk.Header.SessionInfoUpdate;
+                    if (newUpdate != lastUpdate)
+                    {
+                        lastUpdate = newUpdate;
+                        _session = sdk.GetSerializedSessionInfo();
+                    }
 
-                        //if (numBytes > 0)
-                       // {
-                            //File.WriteAllBytes($".\\{Guid.NewGuid()}.ibt", bytes);
-                       // }
+                    if(data != null && _session != null)
+                    {
+                        Console.SetCursorPosition(0,0);
 
-                        //foreach (var car in data.Data.Cars.OrderByDescending(x => x.CarIdxLap).ThenByDescending(x => x.CarIdxLapDistPct))
-                        //{
-                         //   var currentData = _session.DriverInfo.Drivers.Where(y => y.CarIdx == car.CarIdx).FirstOrDefault();
-                          // if (currentData != null && car.CarIdxEstTime != 0)
-                           // {
-                               //Console.WriteLine($"{currentData.CarNumber}\t{string.Format("{0:0.00}", car.CarIdxEstTime)}\t{string.Format("{0:0.00}", car.CarIdxLapDistPct * 100)}");
-                                Console.WriteLine($"SeesionTick: {sdk.Session.SessionTick.ToString()}");
-                            //}
-                        //}
+
+                        foreach (var car in data.Data.Cars.OrderByDescending(x => x.CarIdxLap).ThenByDescending(x => x.CarIdxLapDistPct))
+                        {
+                            var currentData = _session.DriverInfo.Drivers.Where(y => y.CarIdx == car.CarIdx).FirstOrDefault();
+                            if (currentData != null && car.CarIdxEstTime != 0)
+                            {
+                                Console.WriteLine($"{currentData.CarNumber}\t{string.Format("{0:0.00}", car.CarIdxEstTime)}\t{string.Format("{0:0.00}", car.CarIdxLapDistPct * 100)}");
+                            }
+
+                        }
 
                         //foreach (var driver in _session.DriverInfo.Drivers.Where(x => x.IsSpectator == 0).Where(x => x.CarIsPaceCar == "0").Where(x => x.CarIsAI == "0"))
                         //{
@@ -153,7 +152,7 @@ namespace irsdkSharp.ConsoleTest
                         //        var a = "";
                         //    }
                         //}
-                    //}
+                    }
 
                 }
                 else if (_hasConnected)
@@ -171,15 +170,7 @@ namespace irsdkSharp.ConsoleTest
                     _DriverId = -1;
 
                     //Try to find the sim
-                    Console.WriteLine("Connecting to SDK.");
-                    if (sdk.Startup())
-                    {
-                        Console.WriteLine("Connected.");   
-                    }
-                    else
-                    {
-                        Console.WriteLine("Failed to connect.");
-                    }
+                    sdk.Startup();
                 }
 
                 // Sleep for a short amount of time until the next update is available