--- a/src/irsdkSharp.Serialization/irsdkSharp.Serialization.csproj
+++ b/src/irsdkSharp.Serialization/irsdkSharp.Serialization.csproj
@@ -43,7 +43,7 @@
   
   <ItemGroup>
     <PackageReference Include="System.Text.Json" Version="5.0.0" />
-    <PackageReference Include="YamlDotNet" Version="8.1.2" />
+    <PackageReference Include="YamlDotNet" Version="11.2.1" />
   </ItemGroup>
 
   <ItemGroup>