… src/global.json
… src/irsdkSharp.Serialization/Models/Session/IRacingSessionModel.cs
… src/irsdkSharp.sln
… src/irsdkSharp/AccessorExtensions.cs
… src/irsdkSharp/Enums/CarLeftRight.cs
… src/irsdkSharp/Enums/EngineWarnings.cs
… src/irsdkSharp/Enums/PaceFlags.cs
… src/irsdkSharp/Enums/PaceMode.cs
… src/irsdkSharp/Enums/PitServiceFlags.cs
… src/irsdkSharp/Enums/PitServiceStatus.cs
… src/irsdkSharp/Enums/SessionFlags.cs
… src/irsdkSharp/Enums/TrackSurface.cs
… src/irsdkSharp/Enums/TrackSurfaceMaterial.cs
… src/irsdkSharp/Models/CameraInfo/CameraInfoModel.cs
… src/irsdkSharp/Models/CameraInfo/CameraModel.cs
… src/irsdkSharp/Models/CameraInfo/GroupModel.cs
… src/irsdkSharp/Models/Data.cs
… src/irsdkSharp/Models/DriverInfo/DriverInfoModel.cs
… src/irsdkSharp/Models/DriverInfo/DriverModel.cs
… src/irsdkSharp/Models/IRacingSdkHeader.cs
… src/irsdkSharp/Models/QualifyResultsInfo/QualifyResultsInfoModel.cs
… src/irsdkSharp/Models/QualifyResultsInfo/ResultModel.cs
… src/irsdkSharp/Models/RadioInfo/FrequencyModel.cs
… src/irsdkSharp/Models/RadioInfo/RadioInfoModel.cs
… src/irsdkSharp/Models/RadioInfo/RadioModel.cs
… src/irsdkSharp/Models/Session.cs
… src/irsdkSharp/Models/SessionInfo/FastestLapModel.cs
… src/irsdkSharp/Models/SessionInfo/PositionModel.cs
… src/irsdkSharp/Models/SessionInfo/SessionInfoModel.cs
… src/irsdkSharp/Models/SessionInfo/SessionModel.cs