# Summary

This PR includes some refactoring primarily for performance improvements.

Changes include: 
- Refactored Startup to make opening the WaitHandle optional for testing purposes
- Refactored `GetData` resulting in a ~5x speed improvement and allocating about 20x less
- Created `Session` model which allows for real-time just in time de-serialization of session data
  - Improved model by adding more enums 
- Added tests using a serialized ibt file
- Introduced `ReadString` extension to `MemoryMappedViewAccessor`
  -  This improved perf of GetSessionInfo -  10x faster, allocations 20x less)

Benchmark for old vs new `GetData` for a simple float int
```
|     Method |       Mean |    Error |   StdDev | Ratio |  Gen 0 | Allocated |
|----------- |-----------:|---------:|---------:|------:|-------:|----------:|
|    GetData |   242.2 ns |  4.34 ns |  4.06 ns |  0.17 | 0.0057 |      24 B |
| GetDataOld | 1,407.0 ns | 17.08 ns | 15.98 ns |  1.00 | 0.1354 |     568 B |
```

`GetPosition` improvements using `GetData` improvements and `Session` model.
```
|          Method |      Mean |     Error |    StdDev | Ratio |  Gen 0 | Allocated |
|---------------- |----------:|----------:|----------:|------:|-------:|----------:|
|    GetPositions | 30.291 μs | 0.6044 μs | 0.7859 μs |  1.00 | 5.3101 |     22 KB |
| GetPositionsNew |  9.315 μs | 0.1622 μs | 0.1437 μs |  0.30 | 1.7395 |      7 KB |
```

```
|            Method |       Mean |    Error |  StdDev | Ratio |    Gen 0 |    Gen 1 |    Gen 2 | Allocated |
|------------------ |-----------:|---------:|--------:|------:|---------:|---------:|---------:|----------:|
| GetSessionInfoOld | 3,797.6 μs | 10.07 μs | 9.42 μs |  1.00 | 398.4375 | 394.5313 | 394.5313 |  1,570 KB |
|    GetSessionInfo |   225.0 μs |  4.18 μs | 3.91 μs |  0.06 |   8.3008 |   0.7324 |        - |     68 KB |
```