@LuckyNoS7evin 2021-09-09
Okay, so mostly fantastic, however, irsdkSharp.Serialization is where the models are and serialization happens. 
I would love to keep the central lib as close to the c++ version and then have the other bits as add-on.

We do have Data and Session models which would be prime for adjustment.