@christothes 2021-09-10
> Okay, so mostly fantastic, however, irsdkSharp.Serialization is where the models are and serialization happens.
> I would love to keep the central lib as close to the c++ version and then have the other bits as add-on.
> 
> We do have Data and Session models which would be prime for adjustment.

I realize this is a change to the structure, but allow me to explain why I think this is better.

With the old serialization logic, this separation made total sense, because we were taking an all or nothing approach to hydrating the models. The only way to efficiently grab individual data properties was to use the `GetData` method on the core SDK object. Because getting the strongly typed model was expensive, you'd only want to do it for certain scenarios.

In the new `Data` class, using the strongly typed model is much more efficient because we never serialize more than the specific property getter requests. It's even more efficient than `GetData` because we also never need to box to object or perform any logic to determine which type to make the response. Using the strongly typed model is now both the most efficient and the most convenient way to get any of the telemetry properties. Given this, is there ever a reason someone wouldn't prefer to use the model?

Regarding the design being closer to the C++ SDK, I don't think anyone would mind. Most C# developers would prefer to work with objects over requesting properties by name or lookup.

All that said, I respect that this is your project and I won't be offended if you disagree. I just think that with these changes, separating out serialization logic makes things more complicated than they need to be for arguably no real advantage other than a slightly smaller core package. If the dependency on YamlDotNet is a concern, I think we could eliminate the need for it given how predictable the yaml is for Session data. 

Thoughts?