@christothes 2021-09-10
BTW, this PR can be closed in favor of #37 (although w can continue the discussion here for context)